% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idf.R
\name{calculateIdf}
\alias{calculateIdf}
\title{Returns list of IDF data generated from provided raster data}
\usage{
calculateIdf(
  x,
  rp = c(2, 5, 10, 25, 50, 100),
  dur = c(1, 2, 5, 10, 24, 48),
  aggfun = "mean",
  dist = "gev",
  ...
)
}
\arguments{
\item{x}{Raster data}

\item{rp}{list of time periods in years}

\item{dur}{list of time durations in hours}

\item{aggfun}{selection method string}

\item{dist}{distribution method string}
}
\value{
list of nc objects
}
\description{
Returns list of IDF data generated from provided raster data
}
\examples{
calculateIdf(rastrData, rp = c(2, 5, 10, 25, 50, 100), dur = c(1, 2, 5, 10, 24, 48), aggfun = "mean", dist = "gev")
}
