% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{ncRead}
\alias{ncRead}
\title{Returns list of nc objects}
\usage{
ncRead(folderPath, ids)
}
\arguments{
\item{folderPath}{A path to folder containing .nc files}

\item{ids}{IDs to be used to select .nc data}
}
\value{
list of nc objects
}
\description{
Returns list of nc objects
}
\examples{
ncRead("C:/folderPath", cellIDs = c(296, 263, 264, 265, 295, 297, 327, 328, 329))
}
