% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCtoIDF.R
\name{NCtoIDF}
\alias{NCtoIDF}
\title{Generates and plots IDF curves for provided .nc data}
\usage{
NCtoIDF(
  folderPath,
  randomValues = FALSE,
  cellIDs = c(296, 263, 264, 265, 295, 297, 327, 328, 329)
)
}
\arguments{
\item{folderPath}{A path to folder containing .nc files}

\item{randomValues}{Toogle random cellIDs}

\item{cellIDs}{IDs to be used to select .nc data}
}
\description{
Generates and plots IDF curves for provided .nc data
}
\examples{
NCtoIDF("C:/folderPath", cellIDs = c(296, 263, 264, 265, 295, 297, 327, 328, 329))
}
